fix(cargo): the .crates.toml commit was never read back (Refs #345)
#346 made the `_fj_register` merge entry-aware, so forjar no longer SCRAMBLES
`$CARGO_HOME/.crates.toml`. That closed the half of #345 forjar caused. It did
not close the half the issue said mattered most:

  "At minimum, validate after writing ... forjar corrupted this file and
   reported converged for the write; the read-back is what was missing."

A correct merge INTO wreckage is still wreckage. Every host that ran a
pre-#346 forjar has a `.crates.toml` cargo already rejects, and cargo rejects
the WHOLE file for one bad entry. `mv -f` cannot fail on content, so
`_fj_register` returned 0, the package resource reported CONVERGED, and
`cargo install --list` went on naming nothing while every binary on the box
kept running. package_check read that empty list and said `missing:<crate>` --
the exact symptom on intel, where sixteen CI runners share one $HOME.

Reproduced on the pre-fix code by replaying the wreckage quoted in the issue
(orphaned array body after `cross`, unclosed `kani-verifier` key) and
registering an unrelated crate: merge exit 0, destination mutated, file still
unparseable.

`_fj_register` now gates the commit on `_fj_crates_ok`, which copies the
candidate into a throwaway CARGO_HOME and runs `cargo install --list` against
it. Ask CARGO, not a TOML library: cargo is the only consumer that matters and
it is the parser that rejected the file. On refusal the temp file is removed,
the destination is left byte-identical, the operator is told which file and
whether it was already broken before this run, and `return 1` under
`set -euo pipefail` FAILS the resource instead of lying about it.

The probe is fail-open on absent cargo, failed mktemp and failed cp, so a
broken /tmp cannot wedge every install on a host. It costs 0.015s per
registered crate and needs no network.

BEHAVIOUR CHANGE, DELIBERATELY. A machine whose `.crates.toml` is already
wreckage now fails its cargo package resources loudly instead of appending to a
file cargo cannot read. The fix is to repair or move that file aside, not to
revert this. On the cache-hit path the binaries are installed before
registration is refused, so such a host ends up with working binaries cargo
does not know about -- and the check then honestly reports them missing.

Concurrency is untouched and out of scope: sixteen runners sharing one $HOME
can still lose an entry through read-merge-mv. That is a lost entry, not a
corrupt file (#331, #320).

FILE LAYOUT. package.rs was 567 lines, over the 500-line limit and not on the
exception list, and this adds ~30 more. It becomes a module directory --
`package/mod.rs` (apt, uv, brew) and `package/cargo.rs` (the whole cargo
provider: install, cache, registration, removal, state query) -- matching
backup_sync/, nas_archive/ and the other split resources. Two shared helpers,
`per_package_script` and `per_package_query`, replace the seven hand-rolled
copies of "one generated line per package"; the emitted shell is unchanged,
which the existing 13,128 lib tests hold.

tests/falsification_crates_toml_merge.rs gains
`a_merge_into_an_unparseable_registry_is_refused_not_committed`, which runs the
generated shell over the issue's wreckage and asserts the apply fails, the
destination is byte-identical, no `.forjar.*` temp survives, and stderr names
the file. Verified to fail with the gate removed and pass with it. Its harness
is now name-agnostic (`shell_prelude` collects every `_fj_*` function) so the
tests measure behaviour, not the presence of a helper.

The existing fixtures used `(registry+x)` as the source. cargo validates that
URL, so they now use the real crates.io index URL -- a fake that only had to
satisfy a TOML parser no longer satisfies the parser we ask.
